Move the CloudWatch log backend into rise-backend-ecs
logs/cloudwatch.rs is the ECS runtime-log reader — 1,159 lines that stayed in
the permissively-licensed root crate when the ECS backend was relicensed,
leaving ECS code on the wrong side of the boundary.

It moves as-is. Its only root-crate dependency was crate::db::models, itself a
rise-backend-core re-export, and everything else it used already lives in
rise-backend-core::logs, so this is the pure move that split predicted.
EcsCloudWatchContext goes with it: the AWS context is what the ECS controller
hands its own log reader, and both ends now sit in the same crate.

aws-sdk-cloudwatchlogs drops out of rise-deploy's dependencies — nothing in
src/ referenced it any more, and the ECS crate brings its own.

The whole suite is unchanged at 1,515 passing; twelve of them now run in the
ECS crate, which goes from 70 tests to 82.
