+++ Cargo.toml
@@ -137,7 +137,6 @@ backend = [
     "dep:aws-config",
     "dep:aws-sdk-ecr",
     "dep:aws-sdk-ecs",
-    "dep:aws-sdk-cloudwatchlogs",
     "dep:aws-sdk-ssm",
     "dep:aws-sdk-iam",
     "dep:aws-sdk-kms",
@@ -224,7 +223,6 @@ tera = { version = "1.20", optional = true }
 aws-config = { version = "1.5", features = ["behavior-version-latest"], optional = true }
 aws-sdk-ecr = { version = "1.80", default-features = false, features = ["default-https-client", "rt-tokio"], optional = true }
 aws-sdk-ecs = { version = "1.80", default-features = false, features = ["default-https-client", "rt-tokio"], optional = true }
-aws-sdk-cloudwatchlogs = { version = "1.80", default-features = false, features = ["default-https-client", "rt-tokio"], optional = true }
 aws-sdk-ssm = { version = "1.80", default-features = false, features = ["default-https-client", "rt-tokio"], optional = true }
 aws-sdk-iam = { version = "1.80", default-features = false, features = ["default-https-client", "rt-tokio"], optional = true }
 aws-sdk-kms = { version = "1.80", default-features = false, features = ["default-https-client", "rt-tokio"], optional = true }
